python_requirements.txt
python-dotenv
google-cloud-storage